+++ packages/uba/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- `uba_user`: building `fullName` depending on the user language (surname at the beginning or at the end)
+  in case it didn't come from client side
 
 ### Changed